Cast CallMethod's size_t parameter, requested, to jint
requested variable's max value is fCapacity. fCapacity is
set to jsize value returned by GetArrayLength(). As jsize
is typedefed to jint, fCapacity and hence requested variable
can fit in jint.

This cast to jint will avoid problems caused by automatic
type promotion of parameters when passed to a variadic function.

+++ core/jni/android/graphics/CreateJavaOutputStreamAdaptor.cpp
@@ -68,7 +68,7 @@ private:
                 requested = fCapacity;
 
             jint n = env->CallIntMethod(fJavaInputStream,
-                                        gInputStream_readMethodID, fJavaByteArray, 0, requested);
+                                        gInputStream_readMethodID, fJavaByteArray, 0, (jint)requested);
             if (env->ExceptionCheck()) {
                 env->ExceptionDescribe();
                 env->ExceptionClear();
@@ -194,7 +194,7 @@ public:
             }
 
             fEnv->CallVoidMethod(fJavaOutputStream, gOutputStream_writeMethodID,
-                                 storage, 0, requested);
+                                 storage, 0, (jint)requested);
             if (env->ExceptionCheck()) {
                 env->ExceptionDescribe();
                 env->ExceptionClear();
